make traffic modes extensible
The old method for declaring traffic modes required stuffing an
unbounded amount of stuff into generator_commands, which is messy and
hard to maintain. This patch pulls traffic modes into their own module.
This cleans up generator_commands and has the added benefit of easy
contribution of different types of traffic.
