+++ generator/common.py
@@ -61,7 +61,6 @@ class TrafficSpec(object):
         return '\n'.join(lines)
 
     def __str__(self):
-        ret = ''
         attrs = [
             ('loss_rate', lambda x: str(x) if x else 'disabled'),
             ('pps', lambda x: str(x) if x else '<= line rate'),
@@ -78,78 +77,12 @@ class TrafficSpec(object):
         return self.__str__()
 
 
-class UdpSpec(TrafficSpec):
-    def __init__(self, pkt_size=60, num_flows=1, imix=False, **kwargs):
-        self.pkt_size = pkt_size
-        self.num_flows = num_flows
-        self.imix = imix
-        super(UdpSpec, self).__init__(**kwargs)
-
-    def __str__(self):
-        s = super(UdpSpec, self).__str__() + '\n'
-        attrs = [
-            ('pkt_size', lambda x: str(x)),
-            ('num_flows', lambda x: str(x)),
-            ('imix', lambda x: 'enabled' if x else 'disabled')
-        ]
-        return s + self._attrs_to_str(attrs, 25)
-
-    def __repr__(self):
-        return self.__str__()
-
-
-class HttpSpec(TrafficSpec):
-    def __init__(self, num_flows=4000, src_port=1001, **kwargs):
-        self.num_flows = num_flows
-        self.src_port = src_port
-        super(HttpSpec, self).__init__(**kwargs)
-
-    def __str__(self):
-        s = super(HttpSpec, self).__str__() + '\n'
-        attrs = [
-            ('num_flows', lambda x: str(x)),
-            ('src_port', lambda x: str(x)),
-        ]
-        return s + self._attrs_to_str(attrs, 25)
-
-    def __repr__(self):
-        return self.__str__()
-
-
-class FlowGenSpec(TrafficSpec):
-    def __init__(self, pkt_size=60, num_flows=10, flow_duration=5,
-                 flow_rate=None, arrival='uniform', duration='uniform',
-                 src_port=1001, **kwargs):
-        self.pkt_size = pkt_size
-        self.num_flows = num_flows
-        self.flow_duration = flow_duration
-        self.flow_rate = flow_rate
-        self.arrival = arrival
-        self.duration = duration
-        self.src_port = src_port
-        super(FlowGenSpec, self).__init__(**kwargs)
-
-    def __str__(self):
-        s = super(FlowGenSpec, self).__str__() + '\n'
-        attrs = [
-            ('pkt_size', lambda x: str(x)),
-            ('num_flows', lambda x: str(x)),
-            ('flow_duration', lambda x: str(x) if x else 'auto'),
-            ('arrival', lambda x: str(x)),
-            ('duration', lambda x: str(x)),
-            ('src_port', lambda x: str(x))
-        ]
-        return s + self._attrs_to_str(attrs, 25)
-
-    def __repr__(self):
-        return self.__str__()
-
-
 class Session(object):
-    def __init__(self, port, spec, tx_pipelines, rx_pipelines):
+    def __init__(self, port, spec, mode, tx_pipelines, rx_pipelines):
         now = time.time()
-        self.__spec = spec
         self.__port = port
+        self.__spec = spec
+        self.__mode = mode
         self.__curr_stats = None
         self.__last_stats = None
         """
@@ -173,6 +106,9 @@ class Session(object):
 
     def spec(self):
         return self.__spec
+
+    def mode(self):
+        return self.__mode
 
     def tx_pipelines(self):
         return self.__tx_pipelines
